# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  controller.AnnouncementCreateRequest:
    properties:
      content:
        type: string
      title:
        type: string
      type:
        type: integer
    type: object
  controller.AnnouncementUpdateRequest:
    properties:
      content:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
    type: object
...
  controller.HolidayCreateRequest:
    properties:
      date:
        type: string
      description:
        type: string
      is_holiday:
        type: integer
      name:
        type: string
    required:
    - date
    - name
    type: object
...
  model.Announcement:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      sort_order:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        description: 响应状态码
        type: integer
      data:
        description: 响应数据，omitempty表示为空时不序列化
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 棋室管理系统后端 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 棋室管理系统 API
  version: "1.0"
paths:
...
  /announcements:
    get:
      consumes:
      - application/json
      description: 获取公告列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Announcement'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取公告列表
      tags:
      - 公告
    post:
      consumes:
      - application/json
      description: 管理员创建新公告
      parameters:
      - description: 公告信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.AnnouncementCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Announcement'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建公告
      tags:
      - 公告
  /announcements/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除公告
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除公告
      tags:
      - 公告
    get:
      consumes:
      - application/json
      description: 根据公告ID获取公告详情
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Announcement'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取公告详情
      tags:
      - 公告
    put:
      consumes:
      - application/json
      description: 管理员更新公告信息
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: string
      - description: 公告更新信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.AnnouncementUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Announcement'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新公告
      tags:
      - 公告
...
  /holidays:
    get:
      consumes:
      - application/json
      description: 获取所有节假日列表
      parameters:
      - description: 是否节假日
        in: query
        name: is_holiday
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取节假日列表
      tags:
      - 节假日
    post:
      consumes:
      - application/json
      description: 创建新的节假日
      parameters:
      - description: 节假日信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.HolidayCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建节假日
      tags:
      - 节假日
  /holidays/{id}:
    delete:
      consumes:
      - application/json
      description: 删除节假日
      parameters:
      - description: 节假日ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除节假日
      tags:
      - 节假日
    get:
      consumes:
      - application/json
      description: 根据ID获取节假日详情
      parameters:
      - description: 节假日ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取节假日详情
      tags:
      - 节假日
    put:
      consumes:
      - application/json
      description: 更新节假日信息
      parameters:
      - description: 节假日ID
        in: path
        name: id
        required: true
        type: string
      - description: 节假日信息
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新节假日
      tags:
      - 节假日
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"